backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
pdfplumber
python-multipart
tabula-py
PyPDF2 